ovf: add ability to set product properties
update the customzie-ova script to take flags for setting a Product
Property field. The key and value can both be set, and the value can be
a string or JSON. The value can be passed via CLI arg, or can be pointed
to a file. Only one property can be set at a time, so if setting
multiple values, call the script multiple times.

If the property is JSON, it may also be useful to set `--pprop-comment`,
which will add a comment to the property with the JSON in text/readable
form, rather than being quoted.